mengubah pengaturan hak akses menjadi per CRUD

+++ routes/web.php
@@ -62,52 +62,95 @@ Route::middleware('auth')->group(function () {
         ]);
     })->name('dashboard');
 
-    Route::middleware(['role:admin,editor', 'permission:manage_roles'])->group(function () {
-        Route::get('/roles', [RoleController::class, 'index'])->name('roles.index');
-        Route::get('/roles/create', [RoleController::class, 'create'])->name('roles.create');
-        Route::post('/roles', [RoleController::class, 'store'])->name('roles.store');
-        Route::get('/roles/{id}/edit', [RoleController::class, 'edit'])->name('roles.edit');
-        Route::put('/roles/{id}', [RoleController::class, 'update'])->name('roles.update');
-        Route::delete('/roles/{id}', [RoleController::class, 'destroy'])->name('roles.destroy');
+    Route::middleware('role:admin,editor')->group(function () {
+        Route::get('/roles', [RoleController::class, 'index'])->middleware('permission:roles.view')->name('roles.index');
+        Route::get('/roles/create', [RoleController::class, 'create'])->middleware('permission:roles.create')->name('roles.create');
+        Route::post('/roles', [RoleController::class, 'store'])->middleware('permission:roles.create')->name('roles.store');
+        Route::get('/roles/{id}/edit', [RoleController::class, 'edit'])->middleware('permission:roles.update')->name('roles.edit');
+        Route::put('/roles/{id}', [RoleController::class, 'update'])->middleware('permission:roles.update')->name('roles.update');
+        Route::delete('/roles/{id}', [RoleController::class, 'destroy'])->middleware('permission:roles.delete')->name('roles.destroy');
     });
 
-    Route::middleware('permission:manage_mahasiswa')->group(function () {
+    Route::middleware('permission:mahasiswa.create')->group(function () {
         Route::get('Create-mahasiswa', [MahasiswaController::class, 'create'])
             ->name('Create-mahasiswa');
 
         Route::post('simpan-mahasiswa', [MahasiswaController::class, 'store'])
             ->name('simpan-mahasiswa');
+    });
 
+    Route::middleware('permission:mahasiswa.view')->group(function () {
         Route::get('Data-mahasiswa', [MahasiswaController::class, 'index'])
             ->name('Data-mahasiswa');
+    });
 
+    Route::middleware('permission:mahasiswa.update')->group(function () {
         Route::get('edit-mahasiswa/{id}', [MahasiswaController::class, 'edit'])
             ->name('edit-mahasiswa');
 
         Route::put('edit-mahasiswa/{id}', [MahasiswaController::class, 'update'])
             ->name('update-mahasiswa');
+    });
 
+    Route::middleware('permission:mahasiswa.delete')->group(function () {
         Route::delete('hapus-mahasiswa/{id}', [MahasiswaController::class, 'destroy'])
             ->name('hapus-mahasiswa');
     });
 
-    Route::resource('buku', BukuController::class)->except(['show'])->middleware('permission:manage_buku');
-    Route::resource('dosen', DosenController::class)->except(['show'])->middleware('permission:manage_dosen');
-    Route::resource('ruangan', RuanganController::class)->except(['show'])->middleware('permission:manage_ruangan');
-    Route::resource('mata-kuliah', MataKuliahController::class)->except(['show'])->middleware('permission:manage_mata_kuliah');
-    Route::resource('transaksi-krs', TransaksiKrsController::class)->except(['show'])->middleware('permission:manage_krs');
+    Route::resource('buku', BukuController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:buku.view')
+        ->middlewareFor(['create', 'store'], 'permission:buku.create')
+        ->middlewareFor(['edit', 'update'], 'permission:buku.update')
+        ->middlewareFor('destroy', 'permission:buku.delete');
+    Route::resource('dosen', DosenController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:dosen.view')
+        ->middlewareFor(['create', 'store'], 'permission:dosen.create')
+        ->middlewareFor(['edit', 'update'], 'permission:dosen.update')
+        ->middlewareFor('destroy', 'permission:dosen.delete');
+    Route::resource('ruangan', RuanganController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:ruangan.view')
+        ->middlewareFor(['create', 'store'], 'permission:ruangan.create')
+        ->middlewareFor(['edit', 'update'], 'permission:ruangan.update')
+        ->middlewareFor('destroy', 'permission:ruangan.delete');
+    Route::resource('mata-kuliah', MataKuliahController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:mata_kuliah.view')
+        ->middlewareFor(['create', 'store'], 'permission:mata_kuliah.create')
+        ->middlewareFor(['edit', 'update'], 'permission:mata_kuliah.update')
+        ->middlewareFor('destroy', 'permission:mata_kuliah.delete');
+    Route::resource('transaksi-krs', TransaksiKrsController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:krs.view')
+        ->middlewareFor(['create', 'store'], 'permission:krs.create')
+        ->middlewareFor(['edit', 'update'], 'permission:krs.update')
+        ->middlewareFor('destroy', 'permission:krs.delete');
     Route::resource('jadwal-perkuliahan', TransaksiJadwalPerkuliahanController::class)
         ->parameters(['jadwal-perkuliahan' => 'jadwalPerkuliahan'])
         ->except(['show'])
-        ->middleware('permission:manage_jadwal_perkuliahan');
+        ->middlewareFor('index', 'permission:jadwal_perkuliahan.view')
+        ->middlewareFor(['create', 'store'], 'permission:jadwal_perkuliahan.create')
+        ->middlewareFor(['edit', 'update'], 'permission:jadwal_perkuliahan.update')
+        ->middlewareFor('destroy', 'permission:jadwal_perkuliahan.delete');
     Route::resource('presensi-perkuliahan', TransaksiPresensiPerkuliahanController::class)
         ->parameters(['presensi-perkuliahan' => 'presensiPerkuliahan'])
         ->except(['show'])
-        ->middleware('permission:manage_presensi_perkuliahan');
+        ->middlewareFor('index', 'permission:presensi_perkuliahan.view')
+        ->middlewareFor(['create', 'store'], 'permission:presensi_perkuliahan.create')
+        ->middlewareFor(['edit', 'update'], 'permission:presensi_perkuliahan.update')
+        ->middlewareFor('destroy', 'permission:presensi_perkuliahan.delete');
     Route::resource('nilai-perkuliahan', TransaksiNilaiPerkuliahanController::class)
         ->parameters(['nilai-perkuliahan' => 'nilaiPerkuliahan'])
         ->except(['show'])
-        ->middleware('permission:manage_nilai_perkuliahan');
-    Route::resource('pembayaran-ukt', TransaksiPembayaranUktController::class)->except(['show'])->middleware('permission:manage_pembayaran_ukt');
-    Route::resource('peminjaman-buku', TransaksiPeminjamanBukuController::class)->except(['show'])->middleware('permission:manage_peminjaman_buku');
+        ->middlewareFor('index', 'permission:nilai_perkuliahan.view')
+        ->middlewareFor(['create', 'store'], 'permission:nilai_perkuliahan.create')
+        ->middlewareFor(['edit', 'update'], 'permission:nilai_perkuliahan.update')
+        ->middlewareFor('destroy', 'permission:nilai_perkuliahan.delete');
+    Route::resource('pembayaran-ukt', TransaksiPembayaranUktController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:pembayaran_ukt.view')
+        ->middlewareFor(['create', 'store'], 'permission:pembayaran_ukt.create')
+        ->middlewareFor(['edit', 'update'], 'permission:pembayaran_ukt.update')
+        ->middlewareFor('destroy', 'permission:pembayaran_ukt.delete');
+    Route::resource('peminjaman-buku', TransaksiPeminjamanBukuController::class)->except(['show'])
+        ->middlewareFor('index', 'permission:peminjaman_buku.view')
+        ->middlewareFor(['create', 'store'], 'permission:peminjaman_buku.create')
+        ->middlewareFor(['edit', 'update'], 'permission:peminjaman_buku.update')
+        ->middlewareFor('destroy', 'permission:peminjaman_buku.delete');
 });
